Improve save state/load state
1. Move saving and loading state to only occur during VBlank so that we do not have to store screen or line sprite info.
2. Make util class and add ReadBytes/WriteBytes helper
3. Stoping using stream insertion/extraction operators for binary data.

+++ nes/new_ppu.h
@@ -217,7 +217,7 @@ private:
     bool GetBackgroundColor(u8& paletteIndex);
     bool GetSpriteColor(u8 x, u8 y, bool backgroundOpaque, u8& paletteIndex, SpritePriority& priority);
     void ProcessSprites();
-    void PutPixel(u8 x, u8 y, rgb& pixel);
+    void PutPixel(u16 x, u16 y, rgb& pixel);
 
 public:
     u8 Screen[256 * 240 * 3];
@@ -257,7 +257,7 @@ private:
 
     // the current cycle number, 341 cycles in a scan line, 0 - 340
     // TODO: figure out the extra _cycle/odd frame thing.
-    u32 _cycle;
+    u16 _cycle;
 
     // the current scane line number, 
     // 256 pixels wide, 
@@ -271,7 +271,7 @@ private:
     // The current scanline number means that vram write occue "while the scanline is being drawn"
     // This means x scrolls before cycle 256 will take effect on the next scanline (hori(v) = hori(t) at 256 of visible scanlines)
     // Y scrolls effect the next full frame and are not latched in until near the end of _scanline 261
-    u32 _scanline;
+    u16 _scanline;
 
     // Toggled on every frame
     // If true, the last cycle of the pre render scanline is skipped
